+++ transform/src/index.ts
@@ -8,6 +8,7 @@ import { fileURLToPath } from "url";
 import path from "path";
 import fs from "fs";
 import { ThrowReplacer } from "./passes/replacer.js";
+import { StdlibThrowRewriter } from "./passes/stdlib.js";
 
 let WRITE = process.env["WRITE"];
 export default class Transformer extends Transform {
@@ -50,6 +51,8 @@ export default class Transformer extends Transform {
     Globals.baseCWD = path.join(process.cwd(), this.baseDir).replaceAll("\\", "/");
 
     SourceLinker.link(sources);
+
+    StdlibThrowRewriter.rewrite(sources);
 
     ThrowReplacer.replace(sources);